Improved packaging, organization, save and load of models and removed unnecessary initialization
